+++ EveLib.EveOnline/Models/Character/NotificationList.cs
@@ -16,7 +16,7 @@ namespace eZet.EveLib.Modules.Models.Character {
             public long NotificationId { get; set; }
 
             [XmlAttribute("typeID")]
-            public long TypeId { get; set; }
+            public int TypeId { get; set; }
 
             [XmlAttribute("senderID")]
             public long SenderId { get; set; }
